+++ conjur/init/init_controller.py
@@ -38,16 +38,12 @@ class InitController:
         Method that facilitates all method calls in this class
         """
         fetched_certificate = self.get_server_certificate()
+        self.write_certificate(fetched_certificate)
 
-        self.get_account_info(self.conjurrc_data, fetched_certificate)
+        self.get_account_info(self.conjurrc_data)
 
-        self.write_certificate(self.init_logic,
-                               self.conjurrc_data,
-                               self.conjurrc_data.cert_file,
-                               fetched_certificate)
-        self.init_logic.write_conjurrc(conjur.constants.DEFAULT_CONFIG_FILE,
-                                               self.conjurrc_data,
-                                               self.force_overwrite)
+        self.write_conjurrc()
+
         sys.stdout.write("Configuration initialized successfully!\n")
         sys.stdout.write("To begin using the CLI, log in to the Conjur server by" \
                          "running `conjur authn login`")
@@ -57,55 +53,85 @@ class InitController:
         Method to get the certificate from the Conjur endpoint detailed by the user
         """
         # pylint: disable=line-too-long
-        if not self.conjurrc_data.appliance_url:
+        if self.conjurrc_data.appliance_url is None:
             self.conjurrc_data.appliance_url = input("Enter the URL of your Conjur server: ")
-        url = self.conjurrc_data.appliance_url
 
-        self.conjurrc_data.appliance_url = ''.join(url)
         url = urlparse(self.conjurrc_data.appliance_url)
-        # Fetch certificate information if connection is expected over TLS
-        certificate_path = self.conjurrc_data.cert_file if self.conjurrc_data.cert_file else None
-        if certificate_path is not None:
-            self.conjurrc_data.cert_file = ''.join(certificate_path)
-        if certificate_path is None and url.scheme == "https":
-            logging.debug(f"Initiating a TLS Connection with '{self.conjurrc_data.appliance_url}'")
-            fingerprint, fetched_certificate = self.init_logic.get_certificate(url.hostname, url.port)
-            sys.stdout.write(f"\nThe server's certificate SHA-1 fingerprint is:\n{fingerprint}\n")
+        if url.scheme == "http":
+            return None
 
-            sys.stdout.write("\nTo verify this certificate, it is recommended to run the following " \
-                             "command on the Conjur server:\n" \
-                             "openssl x509 -fingerprint -noout -in ~conjur/etc/ssl/conjur.pem\n\n")
-            trust_certificate = input("Trust this certificate? [no]: ")
-            if trust_certificate.lower() != 'yes':
-                raise ValueError("You decided not to trust the certificate")
+        if self.conjurrc_data.cert_file is not None:
+            # Return None because we do not need to fetch the certificate
+            return None
+        # pylint: disable=logging-fstring-interpolation
+        logging.debug(f"Initiating a TLS connection with '{self.conjurrc_data.appliance_url}'")
+        fingerprint, fetched_certificate = self.init_logic.get_certificate(url.hostname, url.port)
 
-            return fetched_certificate
-        return None
+        sys.stdout.write(f"\nThe server's certificate SHA-1 fingerprint is:\n{fingerprint}\n")
+        sys.stdout.write("\nTo verify this certificate, it is recommended to run the following " \
+                         "command on the Conjur server:\n" \
+                         "openssl x509 -fingerprint -noout -in ~conjur/etc/ssl/conjur.pem\n\n")
+        trust_certificate = input("Trust this certificate? [no]: ")
+        if trust_certificate.lower() != 'yes':
+            raise ValueError("You decided not to trust the certificate")
 
-    def get_account_info(self, conjurrc_data, fetched_certificate):
+        return fetched_certificate
+
+    def get_account_info(self, conjurrc_data):
         """
-        Method to fetch the account  from the user
+        Method to fetch the account from the user
         """
-        if not conjurrc_data.account:
-            self.init_logic.fetch_account_from_server(self.conjurrc_data,
-                                                              fetched_certificate)
-        else:
-            conjurrc_data.account = ''.join(conjurrc_data.account)
+        if conjurrc_data.account is None:
+            try:
+                self.init_logic.fetch_account_from_server(self.conjurrc_data)
+            # pylint: disable=broad-except,logging-fstring-interpolation
+            except Exception as error:
+                # pylint: disable=line-too-long,logging-fstring-interpolation
+                logging.debug(f"Unable to fetch the account from the Conjur server. Reason: {error}")
+                # If there was a problem fetching the account from the server, we will request one
+                conjurrc_data.account = input("Enter your organization account name: ")
 
-        if not conjurrc_data.account:
-            raise RuntimeError("Error: account is required")
+                if conjurrc_data.account is None or conjurrc_data.account == '':
+                    # pylint: disable=raise-missing-from
+                    raise RuntimeError("Error: account is required")
 
-    def write_certificate(self, init_logic,
-                          conjurrc_data, certificate_path,
-                          fetched_certificate):
+    def write_certificate(self, fetched_certificate):
         """
         Method to write the certificate fetched from the Conjur endpoint on the user's machine
         """
-        url = urlparse(conjurrc_data.appliance_url)
-        if certificate_path is None and url.scheme == "https":
-            certificate_path = os.path.join(os.path.dirname(conjur.constants.DEFAULT_CONFIG_FILE),
-                                            f"conjur-{conjurrc_data.account}.pem")
-            conjurrc_data.cert_file = certificate_path
-            init_logic.write_certificate_to_file(fetched_certificate,
-                                                 certificate_path,
-                                                 self.force_overwrite)
+        url = urlparse(self.conjurrc_data.appliance_url)
+        if self.conjurrc_data.cert_file is None and url.scheme == "https":
+            # pylint: disable=line-too-long
+            self.conjurrc_data.cert_file = os.path.join(os.path.dirname(conjur.constants.DEFAULT_CONFIG_FILE),
+                                            conjur.constants.CERTIFICATE_FILENAME)
+            is_file_written = self.init_logic.write_certificate_to_file(fetched_certificate,
+                                                                       self.conjurrc_data.cert_file,
+                                                                       self.force_overwrite)
+            if not is_file_written:
+                self.__ensure_overwrite_file(self.conjurrc_data.cert_file)
+                self.init_logic.write_certificate_to_file(fetched_certificate,
+                                                          self.conjurrc_data.cert_file,
+                                                          True)
+
+            sys.stdout.write(f"Wrote certificate to {conjur.constants.CERTIFICATE_FILENAME}\n\n")
+
+    def write_conjurrc(self):
+        """
+        Method to write the conjurrc configuration on the user's machine
+        """
+        is_file_written = self.init_logic.write_conjurrc(conjur.constants.DEFAULT_CONFIG_FILE,
+                                               self.conjurrc_data,
+                                               self.force_overwrite)
+        if not is_file_written:
+            self.__ensure_overwrite_file(conjur.constants.DEFAULT_CONFIG_FILE)
+            self.init_logic.write_conjurrc(conjur.constants.DEFAULT_CONFIG_FILE,
+                                           self.conjurrc_data,
+                                           True)
+        sys.stdout.write(f"Wrote configuration to {conjur.constants.DEFAULT_CONFIG_FILE}\n\n")
+
+    @staticmethod
+    def __ensure_overwrite_file(config_file):
+        force_overwrite = input(f"File {config_file} exists. " \
+                                f"Overwrite [yes]: ")
+        if force_overwrite != '' and force_overwrite.lower() != 'yes':
+            raise Exception(f"Not overwriting {config_file}")